refactor - changed SignForm elements
